test: move algorithm tests to separate classes, add test cases for finding shortest path
+ make internal Dijkstra and Ford-Bellman private
