+++ app/src/test/kotlin/model/DirectedGraphTest.kt
@@ -436,527 +436,4 @@ class DirectedGraphTest {
             }
         }
     }
-
-    @Nested
-    inner class FindSCCTest {
-        @Nested
-        inner class `SCC should return not empty array` {
-            @TestAllDirectedGraphs
-            fun `graph has two connected vertices`(graph: DirectedGraph<Int>) {
-                val v1 = graph.addVertex(1)
-                val v2 = graph.addVertex(2)
-
-                graph.addEdge(v1, v2)
-                graph.addEdge(v2, v1)
-
-                val actualValue = graph.findSCC()
-                val expectedValue = mutableSetOf(mutableSetOf(v1, v2))
-
-                assertEquals(expectedValue, actualValue)
-            }
-
-            @TestAllDirectedGraphs
-            fun `complex graph`(graph: DirectedGraph<Int>) {
-                val v1 = graph.addVertex(1)
-                val v2 = graph.addVertex(2)
-                val v3 = graph.addVertex(3)
-                val v4 = graph.addVertex(4)
-
-                graph.apply {
-                    addEdge(v1, v2)
-                    addEdge(v2, v3)
-                    addEdge(v3, v1)
-                    addEdge(v3, v4)
-                }
-
-                val actualValue = graph.findSCC()
-                val expectedValue = mutableSetOf(mutableSetOf(v1, v2, v3), mutableSetOf(v4))
-
-                assertEquals(expectedValue, actualValue)
-            }
-
-            @TestAllDirectedGraphs
-            fun `graph has multiple SCCs`(graph: DirectedGraph<Int>) {
-                val v1 = graph.addVertex(1)
-                val v2 = graph.addVertex(2)
-                val v3 = graph.addVertex(3)
-                val v4 = graph.addVertex(4)
-                val v5 = graph.addVertex(5)
-
-                graph.apply {
-                    addEdge(v1, v2)
-                    addEdge(v2, v1)
-                    addEdge(v3, v4)
-                    addEdge(v4, v3)
-                    addEdge(v5, v1)
-                }
-
-                val actualValue = graph.findSCC()
-                val expectedValue = mutableSetOf(mutableSetOf(v3, v4), mutableSetOf(v1, v2), mutableSetOf(v5))
-
-                assertEquals(expectedValue, actualValue)
-            }
-
-            @TestAllDirectedGraphs
-            fun `graph with nested cycles`(graph: DirectedGraph<Int>) {
-                val v1 = graph.addVertex(1)
-                val v2 = graph.addVertex(2)
-                val v3 = graph.addVertex(3)
-                val v4 = graph.addVertex(4)
-                val v5 = graph.addVertex(5)
-                val v6 = graph.addVertex(6)
-
-                graph.apply {
-                    addEdge(v1, v2)
-                    addEdge(v2, v3)
-                    addEdge(v3, v1)
-                    addEdge(v3, v4)
-                    addEdge(v4, v5)
-                    addEdge(v5, v6)
-                    addEdge(v6, v4)
-                }
-
-                val actualValue = graph.findSCC()
-                val expectedValue = mutableSetOf(mutableSetOf(v1, v2, v3), mutableSetOf(v4, v5, v6))
-
-                assertEquals(expectedValue, actualValue)
-            }
-
-            @TestAllDirectedGraphs
-            fun `graph with cross connections`(graph: DirectedGraph<Int>) {
-                val v1 = graph.addVertex(1)
-                val v2 = graph.addVertex(2)
-                val v3 = graph.addVertex(3)
-                val v4 = graph.addVertex(4)
-                val v5 = graph.addVertex(5)
-                val v6 = graph.addVertex(6)
-
-                graph.apply {
-                    addEdge(v1, v2)
-                    addEdge(v2, v3)
-                    addEdge(v3, v1)
-                    addEdge(v3, v4)
-                    addEdge(v4, v5)
-                    addEdge(v5, v6)
-                    addEdge(v6, v4)
-                }
-
-                val actualValue = graph.findSCC()
-                val expectedValue = mutableSetOf(mutableSetOf(v1, v2, v3), mutableSetOf(v4, v5, v6))
-
-                assertEquals(expectedValue, actualValue)
-            }
-
-            @TestAllDirectedGraphs
-            fun `graph with disconnected subgraphs`(graph: DirectedGraph<Int>) {
-                val v1 = graph.addVertex(1)
-                val v2 = graph.addVertex(2)
-                val v3 = graph.addVertex(3)
-                val v4 = graph.addVertex(4)
-                val v5 = graph.addVertex(5)
-                val v6 = graph.addVertex(6)
-
-                graph.apply {
-                    addEdge(v1, v2)
-                    addEdge(v2, v1)
-                    addEdge(v3, v4)
-                    addEdge(v4, v3)
-                    addEdge(v5, v6)
-                    addEdge(v6, v5)
-                }
-
-                val actualValue = graph.findSCC()
-                val expectedValue = mutableSetOf(mutableSetOf(v1, v2), mutableSetOf(v3, v4), mutableSetOf(v5, v6))
-                assertEquals(expectedValue, actualValue)
-            }
-
-            @Disabled("Our model doesn't support edge from vertex to itself, check DirectedGraph.kt")
-            @TestAllDirectedGraphs
-            fun `graph with single vertex cycle`(graph: DirectedGraph<Int>) {
-                val v1 = graph.addVertex(1)
-                val v2 = graph.addVertex(2)
-                val v3 = graph.addVertex(3)
-
-                graph.apply {
-                    addEdge(v1, v2)
-                    addEdge(v2, v3)
-                    addEdge(v3, v1)
-                    addEdge(v3, v3)
-                }
-
-                val actualValue = graph.findSCC()
-                val expectedValue = mutableSetOf(mutableSetOf(v1, v2, v3))
-
-                assertEquals(expectedValue, actualValue)
-            }
-        }
-
-        @Nested
-        inner class `SCC should return single-element SCCs` {
-            @TestAllDirectedGraphs
-            fun `graph has single vertex`(graph: DirectedGraph<Int>) {
-                val v1 = graph.addVertex(1)
-
-                val actualValue = graph.findSCC()
-                val expectedValue = mutableSetOf(mutableSetOf(v1))
-
-                assertEquals(expectedValue, actualValue)
-            }
-
-            @TestAllDirectedGraphs
-            fun `graph with multiple disconnected vertices`(graph: DirectedGraph<Int>) {
-                val v1 = graph.addVertex(1)
-                val v2 = graph.addVertex(2)
-                val v3 = graph.addVertex(3)
-                val v4 = graph.addVertex(4)
-
-                val actualValue = graph.findSCC()
-                val expectedValue =
-                    mutableSetOf(
-                        mutableSetOf(v1),
-                        mutableSetOf(v2),
-                        mutableSetOf(v3),
-                        mutableSetOf(v4)
-                    )
-
-                assertEquals(expectedValue, actualValue)
-            }
-        }
-
-        @Nested
-        inner class `Additional edge cases`() {
-            @TestAllDirectedGraphs
-            fun `empty graph`(graph: DirectedGraph<Int>) {
-
-                val actualValue = graph.findSCC()
-                val expectedValue = mutableSetOf<MutableSet<Vertex<Int>>>()
-
-                assertEquals(expectedValue, actualValue)
-            }
-
-            @Disabled("Our model doesn't support edge from vertex to itself, check DirectedGraph.kt")
-            @TestAllDirectedGraphs
-            fun `graph with self-loops`(graph: DirectedGraph<Int>) {
-                val v1 = graph.addVertex(1)
-                val v2 = graph.addVertex(2)
-
-                graph.addEdge(v1, v1)
-                graph.addEdge(v2, v2)
-
-                val actualValue = graph.findSCC()
-                val expectedValue = mutableSetOf(mutableSetOf(v1), mutableSetOf(v2))
-
-                assertEquals(expectedValue, actualValue)
-            }
-
-            @TestAllDirectedGraphs
-            fun `linear graph`(graph: DirectedGraph<Int>) {
-                val v1 = graph.addVertex(1)
-                val v2 = graph.addVertex(2)
-                val v3 = graph.addVertex(3)
-
-                graph.addEdge(v1, v2)
-                graph.addEdge(v2, v3)
-
-                val actualValue = graph.findSCC()
-                val expectedValue = mutableSetOf(mutableSetOf(v3), mutableSetOf(v2), mutableSetOf(v1))
-
-                assertEquals(expectedValue, actualValue)
-            }
-
-            @TestAllDirectedGraphs
-            fun `graph with cycles and tail`(graph: DirectedGraph<Int>) {
-                val v1 = graph.addVertex(1)
-                val v2 = graph.addVertex(2)
-                val v3 = graph.addVertex(3)
-                val v4 = graph.addVertex(4)
-                val v5 = graph.addVertex(5)
-
-                graph.apply {
-                    addEdge(v1, v2)
-                    addEdge(v2, v3)
-                    addEdge(v3, v1)
-                    addEdge(v4, v3)
-                    addEdge(v4, v5)
-                }
-
-                val actualValue = graph.findSCC()
-                val expectedValue = mutableSetOf(mutableSetOf(v1, v2, v3), mutableSetOf(v4), mutableSetOf(v5))
-
-                assertEquals(expectedValue, actualValue)
-            }
-        }
-
-        @Nested
-        inner class `Side-effects check` {
-            @TestAllDirectedGraphs
-            fun `check vertices in complex graph`(graph: DirectedGraph<Int>) {
-                val v1 = graph.addVertex(1)
-                val v2 = graph.addVertex(2)
-                val v3 = graph.addVertex(3)
-                val v4 = graph.addVertex(4)
-
-                graph.apply {
-                    addEdge(v1, v2)
-                    addEdge(v2, v3)
-                    addEdge(v3, v1)
-                    addEdge(v3, v4)
-                }
-
-                val expectedValue = graph.getVertices()
-                graph.findSCC()
-                val actualValue = graph.getVertices()
-
-                assertEquals(expectedValue, actualValue)
-            }
-
-            @TestAllDirectedGraphs
-            fun `check edges in complex graph`(graph: DirectedGraph<Int>) {
-                val v1 = graph.addVertex(1)
-                val v2 = graph.addVertex(2)
-                val v3 = graph.addVertex(3)
-                val v4 = graph.addVertex(4)
-
-                graph.apply {
-                    addEdge(v1, v2)
-                    addEdge(v2, v3)
-                    addEdge(v3, v1)
-                    addEdge(v3, v4)
-                }
-
-                val expectedValue = graph.getEdges()
-                graph.findSCC()
-                val actualValue = graph.getEdges()
-
-                assertEquals(expectedValue, actualValue)
-            }
-
-            @TestAllDirectedGraphs
-            fun `check edges in graph with cycles and tail`(graph: DirectedGraph<Int>) {
-                val v1 = graph.addVertex(1)
-                val v2 = graph.addVertex(2)
-                val v3 = graph.addVertex(3)
-                val v4 = graph.addVertex(4)
-                val v5 = graph.addVertex(5)
-
-                graph.apply {
-                    addEdge(v1, v2)
-                    addEdge(v2, v3)
-                    addEdge(v3, v1)
-                    addEdge(v4, v3)
-                    addEdge(v4, v5)
-                }
-
-                val expectedValue = graph.getEdges()
-                graph.findSCC()
-                val actualValue = graph.getEdges()
-
-                assertEquals(expectedValue, actualValue)
-            }
-
-            @TestAllDirectedGraphs
-            fun `check vertices graph with cycles and tail`(graph: DirectedGraph<Int>) {
-                val v1 = graph.addVertex(1)
-                val v2 = graph.addVertex(2)
-                val v3 = graph.addVertex(3)
-                val v4 = graph.addVertex(4)
-                val v5 = graph.addVertex(5)
-
-                graph.apply {
-                    addEdge(v1, v2)
-                    addEdge(v2, v3)
-                    addEdge(v3, v1)
-                    addEdge(v4, v3)
-                    addEdge(v4, v5)
-                }
-
-                val expectedValue = graph.getVertices()
-                graph.findSCC()
-                val actualValue = graph.getVertices()
-
-                assertEquals(expectedValue, actualValue)
-            }
-        }
-    }
-
-    @Nested
-    inner class FindKeyVerticesTest {
-        @Nested
-        inner class `One vertex is picked over another`() {
-            @TestAllDirectedGraphs
-            fun `if it can reach more vertices`(graph: DirectedGraph<Int>) {
-                val v0 = graph.addVertex(0)
-                val v1 = graph.addVertex(1)
-                val v2 = graph.addVertex(2)
-                val v3 = graph.addVertex(3)
-
-                graph.apply {
-                    addEdge(v0, v1)
-                    addEdge(v0, v2)
-                    addEdge(v0, v3)
-                    addEdge(v1, v2)
-                }
-
-                val expectedResult = setOf(v0)
-                val actualResult = graph.findKeyVertices()
-
-                assertEquals(expectedResult, actualResult)
-            }
-
-            @TestAllDirectedGraphs
-            fun `if it can reach other vertices with fewer edges`(graph: DirectedGraph<Int>) {
-                val v0 = graph.addVertex(0)
-                val v1 = graph.addVertex(1)
-                val v2 = graph.addVertex(2)
-                val v3 = graph.addVertex(3)
-
-                graph.apply {
-                    addEdge(v1, v2)
-                    addEdge(v2, v3)
-                    addEdge(v0, v2)
-                    addEdge(v0, v3)
-                }
-
-                val expectedResult = setOf(v0)
-                val actualResult = graph.findKeyVertices()
-
-                assertEquals(expectedResult, actualResult)
-            }
-        }
-    }
-
-    @Nested
-    inner class FindCyclesTest {
-        @Nested
-        inner class `There are some cycles` {
-            @TestAllDirectedGraphs
-            fun `all cycles should be returned`(graph: DirectedGraph<Int>) {
-                val v0 = graph.addVertex(0)
-                val v1 = graph.addVertex(1)
-                val v2 = graph.addVertex(2)
-                val v3 = graph.addVertex(3)
-                val v4 = graph.addVertex(4)
-                val v5 = graph.addVertex(5)
-                val v6 = graph.addVertex(6)
-                val v7 = graph.addVertex(7)
-                val v8 = graph.addVertex(8)
-
-                val e01 = graph.addEdge(v0, v1)
-                val e07 = graph.addEdge(v0, v7)
-                val e04 = graph.addEdge(v0, v4)
-                val e18 = graph.addEdge(v1, v8)
-                val e12 = graph.addEdge(v1, v2)
-                val e20 = graph.addEdge(v2, v0)
-                val e21 = graph.addEdge(v2, v1)
-                val e25 = graph.addEdge(v2, v5)
-                val e23 = graph.addEdge(v2, v3)
-                val e53 = graph.addEdge(v5, v3)
-                val e34 = graph.addEdge(v3, v4)
-                val e41 = graph.addEdge(v4, v1)
-                val e78 = graph.addEdge(v7, v8)
-                val e87 = graph.addEdge(v8, v7)
-
-                val expectedCycle1 = listOf(e12 to v2, e21 to v1)
-                val expectedCycle2 = listOf(e12 to v2, e20 to v0, e01 to v1)
-                val expectedCycle3 = listOf(e12 to v2, e20 to v0, e04 to v4, e41 to v1)
-                val expectedCycle4 = listOf(e12 to v2, e23 to v3, e34 to v4, e41 to v1)
-                val expectedCycle5 = listOf(e12 to v2, e25 to v5, e53 to v3, e34 to v4, e41 to v1)
-
-                val actualValue = graph.findCycles(v1)
-                val expectedValue =
-                    setOf(
-                        expectedCycle1,
-                        expectedCycle2,
-                        expectedCycle3,
-                        expectedCycle4,
-                        expectedCycle5
-                    )
-
-                assertEquals(expectedValue, actualValue)
-            }
-
-            @TestAllDirectedGraphs
-            fun `SCC of 2 vertices should have one cycle`(graph: DirectedGraph<Int>) {
-                val v0 = graph.addVertex(0)
-                val v1 = graph.addVertex(1)
-                val v2 = graph.addVertex(2)
-                val v3 = graph.addVertex(3)
-
-                val e01 = graph.addEdge(v0, v1)
-                val e12 = graph.addEdge(v1, v2)
-                val e21 = graph.addEdge(v2, v1)
-                val e23 = graph.addEdge(v2, v3)
-
-                val actualValue = graph.findCycles(v1)
-                val expectedValue = setOf(listOf(e12 to v2, e21 to v1))
-
-                assertEquals(expectedValue, actualValue)
-            }
-        }
-
-        @Nested
-        inner class `There are no cycles` {
-            @TestAllDirectedGraphs
-            fun `vertex without outgoing edges shouldn't have cycles`(graph: DirectedGraph<Int>) {
-                val v0 = graph.addVertex(0)
-                val v1 = graph.addVertex(1)
-
-                graph.addEdge(v0, v1)
-
-                val actualValue = graph.findCycles(v1)
-                val expectedValue = emptySet<List<Pair<Edge<Int>, Vertex<Int>>>>()
-
-                assertEquals(expectedValue, actualValue)
-            }
-
-            @TestAllDirectedGraphs
-            fun `SCC of 1 vertex shouldn't have cycles`(graph: DirectedGraph<Int>) {
-                val v0 = graph.addVertex(0)
-                val v1 = graph.addVertex(1)
-
-                graph.addEdge(v0, v1)
-
-                val actualValue = graph.findCycles(v0)
-                val expectedValue = emptySet<List<Pair<Edge<Int>, Vertex<Int>>>>()
-
-                assertEquals(expectedValue, actualValue)
-            }
-        }
-
-        @TestAllDirectedGraphs
-        fun `graph shouldn't change`(graph: DirectedGraph<Int>) {
-            val v0 = graph.addVertex(0)
-            val v1 = graph.addVertex(1)
-            val v2 = graph.addVertex(2)
-            val v3 = graph.addVertex(3)
-            val v4 = graph.addVertex(4)
-            val v5 = graph.addVertex(5)
-            val v6 = graph.addVertex(6)
-            val v7 = graph.addVertex(7)
-            val v8 = graph.addVertex(8)
-
-            val e01 = graph.addEdge(v0, v1)
-            val e07 = graph.addEdge(v0, v7)
-            val e04 = graph.addEdge(v0, v4)
-            val e18 = graph.addEdge(v1, v8)
-            val e12 = graph.addEdge(v1, v2)
-            val e20 = graph.addEdge(v2, v0)
-            val e21 = graph.addEdge(v2, v1)
-            val e25 = graph.addEdge(v2, v5)
-            val e23 = graph.addEdge(v2, v3)
-            val e53 = graph.addEdge(v5, v3)
-            val e34 = graph.addEdge(v3, v4)
-            val e41 = graph.addEdge(v4, v1)
-            val e78 = graph.addEdge(v7, v8)
-            val e87 = graph.addEdge(v8, v7)
-
-            val expectedGraph = graph.getVertices() to graph.getEdges().toSet()
-
-            graph.findCycles(v1)
-
-            val actualGraph = graph.getVertices() to graph.getEdges().toSet()
-
-            assertEquals(expectedGraph, actualGraph)
-        }
-    }
 }
