+++ app/src/test/kotlin/model/WeightedUndirectedGraphTest.kt
@@ -102,405 +102,4 @@ class WeightedUndirectedGraphTest {
             }
         }
     }
-
-    @Nested
-    inner class FindShortestPathDijkstraTest {
-        @Nested
-        inner class `Normal path should be returned`() {
-            @Test
-            fun `all is as usual, should return default`() {
-                val v0 = graph.addVertex(0)
-                val v1 = graph.addVertex(1)
-                val v2 = graph.addVertex(2)
-                val v3 = graph.addVertex(3)
-                val v4 = graph.addVertex(4)
-
-                val e0 = graph.addEdge(v0, v1, 10)
-                val e1 = graph.addEdge(v0, v4, 100)
-                val e2 = graph.addEdge(v0, v3, 30)
-                val e3 = graph.addEdge(v1, v2, 2)
-                val e4 = graph.addEdge(v2, v4, 10)
-                val e5 = graph.addEdge(v3, v2, 20)
-                val e6 = graph.addEdge(v3, v4, 60)
-
-                val expectedResult = listOf(e0 to v1, e3 to v2, e4 to v4)
-                val actualResult = graph.findShortestPathDijkstra(v0, v4)
-
-                assertEquals(expectedResult, actualResult)
-            }
-
-            @Test
-            fun `if graph has multiple paths and equal weights`() {
-                val v0 = graph.addVertex(0)
-                val v1 = graph.addVertex(1)
-                val v2 = graph.addVertex(2)
-                val v3 = graph.addVertex(3)
-
-                val e0 = graph.addEdge(v0, v1, 1)
-                val e1 = graph.addEdge(v0, v2, 1)
-                val e2 = graph.addEdge(v1, v3, 1)
-                val e3 = graph.addEdge(v2, v3, 1)
-
-                val expectedResult1 = listOf(e0 to v1, e2 to v3)
-                val expectedResult2 = listOf(e1 to v2, e3 to v3)
-
-                val actualResult = graph.findShortestPathDijkstra(v0, v3)
-
-                assertTrue(actualResult == expectedResult1 || actualResult == expectedResult2)
-            }
-
-            @Test
-            fun `if graph has single edge`() {
-                val v0 = graph.addVertex(0)
-                val v1 = graph.addVertex(1)
-
-                val e0 = graph.addEdge(v0, v1, 5)
-
-                val expectedResult = listOf(e0 to v1)
-                val actualResult = graph.findShortestPathDijkstra(v0, v1)
-
-                assertEquals(expectedResult, actualResult)
-            }
-
-            @Disabled("Dijkstra's algorithm doesn't work with negative weights")
-            @Test
-            fun `if graph has negative weights`() {
-                val v0 = graph.addVertex(0)
-                val v1 = graph.addVertex(1)
-                val v2 = graph.addVertex(2)
-                val v3 = graph.addVertex(3)
-
-                val e0 = graph.addEdge(v0, v1, -1)
-                val e1 = graph.addEdge(v1, v2, -2)
-                val e2 = graph.addEdge(v2, v3, -3)
-
-                val expectedResult = listOf(e0 to v1, e1 to v2, e2 to v3)
-                val actualResult = graph.findShortestPathDijkstra(v0, v3)
-
-                assertEquals(expectedResult, actualResult)
-            }
-
-            @Test
-            fun `graph has multiple equal shortest paths`() {
-                val v0 = graph.addVertex(0)
-                val v1 = graph.addVertex(1)
-                val v2 = graph.addVertex(2)
-                val v3 = graph.addVertex(3)
-                val v4 = graph.addVertex(4)
-
-                val e0 = graph.addEdge(v0, v1, 1)
-                val e1 = graph.addEdge(v0, v2, 1)
-                val e2 = graph.addEdge(v1, v3, 1)
-                val e3 = graph.addEdge(v2, v3, 1)
-                val e4 = graph.addEdge(v3, v4, 1)
-
-                val expectedResult1 = listOf(e0 to v1, e2 to v3, e4 to v4)
-                val expectedResult2 = listOf(e1 to v2, e3 to v3, e4 to v4)
-                val actualResult = graph.findShortestPathDijkstra(v0, v4)
-
-                assertTrue(actualResult == expectedResult1 || actualResult == expectedResult2)
-            }
-
-            @Test
-            fun `if graph has a cycle`() {
-                val graph = WeightedUndirectedGraph<Int>()
-                val v0 = graph.addVertex(0)
-                val v1 = graph.addVertex(1)
-                val v2 = graph.addVertex(2)
-
-                val e0 = graph.addEdge(v0, v1, 1)
-                val e1 = graph.addEdge(v1, v2, 2)
-                val e2 = graph.addEdge(v2, v0, 3)
-
-                val actualResult = graph.findShortestPathDijkstra(v0, v2)
-                val expectedResult = listOf(e2 to v2)
-
-                assertEquals(expectedResult, actualResult)
-            }
-
-            @Test
-            fun `if path is in other way (not how edges were set)`() {
-                val graph = WeightedUndirectedGraph<Int>()
-                val v0 = graph.addVertex(0)
-                val v1 = graph.addVertex(1)
-                val v2 = graph.addVertex(2)
-
-                val e0 = graph.addEdge(v0, v1, 0)
-                val e1 = graph.addEdge(v1, v2, 0)
-
-                val expectedResult = listOf(e1 to v1, e0 to v0)
-                val actualResult = graph.findShortestPathDijkstra(v2, v0)
-
-                assertEquals(expectedResult, actualResult)
-            }
-
-            @Test
-            fun `if all the edges have zero weight in undirected graph`() {
-                val graph = WeightedUndirectedGraph<Int>()
-                val v0 = graph.addVertex(0)
-                val v1 = graph.addVertex(1)
-                val v2 = graph.addVertex(2)
-
-                val e0 = graph.addEdge(v0, v1, 0)
-                val e1 = graph.addEdge(v1, v2, 0)
-
-                val expectedResult = listOf(e0 to v1, e1 to v2)
-                val actualResult = graph.findShortestPathDijkstra(v0, v2)
-
-                assertEquals(expectedResult, actualResult)
-            }
-        }
-
-        @Nested
-        inner class `No path should be returned`() {
-            @Test
-            fun `no path exists in undirected graph`() {
-                val v0 = graph.addVertex(0)
-                val v1 = graph.addVertex(1)
-                val v2 = graph.addVertex(2)
-                val v3 = graph.addVertex(3)
-
-                graph.addEdge(v0, v1, 1)
-                graph.addEdge(v1, v2, 2)
-
-                val actualResult = graph.findShortestPathDijkstra(v0, v3)
-
-                assertNull(actualResult)
-            }
-
-            @Test
-            fun `if start and end vertices are the same`() {
-                val v0 = graph.addVertex(0)
-                val v1 = graph.addVertex(1)
-                val v2 = graph.addVertex(2)
-
-                graph.addEdge(v0, v1, 1)
-                graph.addEdge(v1, v2, 2)
-
-                val actualResult = graph.findShortestPathDijkstra(v0, v0)
-
-                actualResult?.isEmpty()?.let { assertTrue(it) }
-            }
-
-            @Test
-            fun `if graph has single vertex`() {
-                val v0 = graph.addVertex(0)
-
-                val actualResult = graph.findShortestPathDijkstra(v0, v0)
-
-                actualResult?.isEmpty()?.let { assertTrue(it) }
-            }
-        }
-    }
-
-    @Nested
-    inner class FindMinSpanningTreeTest {
-        @Nested
-        inner class `An edge is picked over another`() {
-            @Test
-            fun `if it has lesser weight but both have positive`() {
-                val v0 = graph.addVertex(0)
-                val v1 = graph.addVertex(1)
-
-                val e01Heavy = graph.addEdge(v0, v1, 5)
-                val e01Light = graph.addEdge(v0, v1, 3)
-
-                val expectedReturn = listOf(e01Light)
-                val actualReturn = graph.findMinSpanningTree()
-
-                assertEquals(expectedReturn, actualReturn)
-            }
-
-            @Test
-            fun `if it has lesser weight but both have negative`() {
-                val v0 = graph.addVertex(0)
-                val v1 = graph.addVertex(1)
-
-                val e01Heavy = graph.addEdge(v0, v1, -5)
-                val e01Light = graph.addEdge(v0, v1, -10)
-
-                val expectedReturn = listOf(e01Light)
-                val actualReturn = graph.findMinSpanningTree()
-
-                assertEquals(expectedReturn, actualReturn)
-            }
-
-            @Test
-            fun `if it has zero weight and other has positive`() {
-                val v0 = graph.addVertex(0)
-                val v1 = graph.addVertex(1)
-
-                val e01Pos = graph.addEdge(v0, v1, 5)
-                val e01Zero = graph.addEdge(v0, v1, 0)
-
-                val expectedReturn = listOf(e01Zero)
-                val actualReturn = graph.findMinSpanningTree()
-
-                assertEquals(expectedReturn, actualReturn)
-            }
-
-            @Test
-            fun `if it has negative weight and other has positive or zero`() {
-                val v0 = graph.addVertex(0)
-                val v1 = graph.addVertex(1)
-
-                val e01Pos = graph.addEdge(v0, v1, 5)
-                val e01Zero = graph.addEdge(v0, v1, 0)
-                val e01Neg = graph.addEdge(v0, v1, -5)
-
-                val expectedReturn = listOf(e01Neg)
-                val actualReturn = graph.findMinSpanningTree()
-
-                assertEquals(expectedReturn, actualReturn)
-            }
-        }
-
-        @Nested
-        inner class `An edge is not picked over another`() {
-            @Test
-            fun `if it forms a cycle and has greatest weight in it`() {
-                val v0 = graph.addVertex(0)
-                val v1 = graph.addVertex(1)
-                val v2 = graph.addVertex(2)
-                val v3 = graph.addVertex(3)
-
-                val e01 = graph.addEdge(v0, v1, 1)
-                val e12 = graph.addEdge(v1, v2, 1)
-                val e23 = graph.addEdge(v2, v3, 1)
-
-                val e30 = graph.addEdge(v3, v0, 5)
-
-                val expectedReturn = setOf(e01, e12, e23)
-                val actualReturn = graph.findMinSpanningTree().toSet()
-
-                assertEquals(expectedReturn, actualReturn)
-            }
-        }
-
-        @Nested
-        inner class `All edges should be returned`() {
-            @Test
-            fun `if graph is a tree`() {
-                val v0 = graph.addVertex(0)
-                val v1 = graph.addVertex(1)
-                val v2 = graph.addVertex(2)
-                val v3 = graph.addVertex(3)
-                val v4 = graph.addVertex(4)
-
-                val eva01 = graph.addEdge(v0, v1, 1)
-                val eva02 = graph.addEdge(v0, v2, 10)
-                val e23 = graph.addEdge(v2, v3, 0)
-                val e24 = graph.addEdge(v2, v4, -20)
-
-                val expectedResult = setOf(e24, e23, eva01, eva02)
-                val actualResult = graph.findMinSpanningTree().toSet()
-
-                assertEquals(expectedResult, actualResult)
-            }
-        }
-
-        @Nested
-        inner class `No edge should be returned`() {
-            @Test
-            fun `if graph has no vertices`() {
-                val expectedResult = listOf<Edge<Int>>()
-                val actualResult = graph.findMinSpanningTree()
-
-                assertEquals(expectedResult, actualResult)
-            }
-
-            @Test
-            fun `if graph has no edges`() {
-                graph.addVertex(0)
-                graph.addVertex(1)
-
-                val expectedResult = listOf<Edge<Int>>()
-                val actualResult = graph.findMinSpanningTree()
-
-                assertEquals(expectedResult, actualResult)
-            }
-        }
-    }
-
-    @Nested
-    inner class FindKeyVerticesTest {
-        @Nested
-        inner class `One vertex is picked over another`() {
-            @Test
-            fun `if it can reach more vertices`() {
-                val v0 = graph.addVertex(0)
-                val v1 = graph.addVertex(1)
-                val v2 = graph.addVertex(2)
-                val v3 = graph.addVertex(3)
-
-                graph.apply {
-                    addEdge(v0, v1, 1)
-                    addEdge(v0, v2, 1)
-                    addEdge(v0, v3, 1)
-                    addEdge(v1, v2, 1)
-                }
-
-                val expectedResult = setOf(v0)
-                val actualResult = graph.findKeyVertices()
-
-                assertEquals(expectedResult, actualResult)
-            }
-
-            @Test
-            fun `if it can reach other vertices with fewer edges`() {
-                val v0 = graph.addVertex(0)
-                val v1 = graph.addVertex(1)
-                val v2 = graph.addVertex(2)
-                val v3 = graph.addVertex(3)
-
-                graph.apply {
-                    addEdge(v0, v1)
-                    addEdge(v0, v2)
-                    addEdge(v0, v3)
-                }
-
-                val expectedResult = setOf(v0)
-                val actualResult = graph.findKeyVertices()
-
-                assertEquals(expectedResult, actualResult)
-            }
-
-            @Test
-            fun `if its sum of distances to other vertices is less`() {
-                val v0 = graph.addVertex(0)
-                val v1 = graph.addVertex(1)
-                val v2 = graph.addVertex(2)
-                val v3 = graph.addVertex(3)
-
-                graph.apply {
-                    addEdge(v0, v2, 1)
-                    addEdge(v0, v3, 1)
-                    addEdge(v1, v2, 2)
-                    addEdge(v2, v3, 2)
-                }
-
-                val expectedResult = setOf(v0)
-                val actualResult = graph.findKeyVertices()
-
-                assertEquals(expectedResult, actualResult)
-            }
-        }
-
-        @Nested
-        inner class `Returns null`() {
-            @Test
-            fun `if graph has negative edges`() {
-                val v0 = graph.addVertex(0)
-                val v1 = graph.addVertex(1)
-                val v2 = graph.addVertex(2)
-
-                graph.addEdge(v0, v1, 1)
-                graph.addEdge(v0, v2, -1)
-
-                val actualResult = graph.findKeyVertices()
-
-                assertNull(actualResult)
-            }
-        }
-    }
 }
